# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Canine Adoption API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/dogs": {
      "get": {
        "tags": [
          "dogs"
        ],
        "summary": "List Dogs",
        "operationId": "list_dogs_dogs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DogRead"
                  },
                  "type": "array",
                  "title": "Response List Dogs Dogs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "dogs"
        ],
        "summary": "Create Dog",
        "operationId": "create_dog_dogs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/dogs/{dog_id}": {
      "get": {
        "tags": [
          "dogs"
        ],
        "summary": "Get Dog",
        "operationId": "get_dog_dogs__dog_id__get",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dog Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "dogs"
        ],
        "summary": "Update Dog",
        "operationId": "update_dog_dogs__dog_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dog Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DogUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "dogs"
        ],
        "summary": "Delete Dog",
        "operationId": "delete_dog_dogs__dog_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dog Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dogs/{dog_id}/ai/overview": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Stream Dog Overview",
        "operationId": "stream_dog_overview_dogs__dog_id__ai_overview_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dog Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dogs/{dog_id}/ai/chat": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Stream Dog Chat",
        "operationId": "stream_dog_chat_dogs__dog_id__ai_chat_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dog Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DogChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "pattern": "^(system|user|assistant)$",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 32000,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage"
      },
      "DogChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "DogChatRequest"
      },
      "DogCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "title": "Name"
          },
          "breed": {
            "type": "string",
            "maxLength": 120,
            "title": "Breed"
          },
          "age": {
            "type": "integer",
            "maximum": 40.0,
            "minimum": 0.0,
            "title": "Age"
          },
          "sex": {
            "type": "string",
            "maxLength": 32,
            "title": "Sex"
          },
          "size": {
            "type": "string",
            "maxLength": 32,
            "title": "Size"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "photo_url": {
            "type": "string",
            "maxLength": 2048,
            "title": "Photo Url"
          },
          "status": {
            "type": "string",
            "pattern": "^(available|adopted)$",
            "title": "Status",
            "default": "available"
          }
        },
        "type": "object",
        "required": [
          "name",
          "breed",
          "age",
          "sex",
          "size",
          "description",
          "photo_url"
        ],
        "title": "DogCreate"
      },
      "DogRead": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "title": "Name"
          },
          "breed": {
            "type": "string",
            "maxLength": 120,
            "title": "Breed"
          },
          "age": {
            "type": "integer",
            "maximum": 40.0,
            "minimum": 0.0,
            "title": "Age"
          },
          "sex": {
            "type": "string",
            "maxLength": 32,
            "title": "Sex"
          },
          "size": {
            "type": "string",
            "maxLength": 32,
            "title": "Size"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "photo_url": {
            "type": "string",
            "maxLength": 2048,
            "title": "Photo Url"
          },
          "status": {
            "type": "string",
            "pattern": "^(available|adopted)$",
            "title": "Status",
            "default": "available"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "breed",
          "age",
          "sex",
          "size",
          "description",
          "photo_url",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "DogRead"
      },
      "DogUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "breed": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Breed"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 40.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "sex": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Sex"
          },
          "size": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(available|adopted)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "DogUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}